Refactored Messages
